Tell me, why did I use floats again?!

+++ app/src/main/kotlin/com/stefanodacchille/playground/State.kt
@@ -3,7 +3,7 @@ package com.stefanodacchille.playground
 open class State() {
 
   data class Init(val displayNumber: DisplayNumber) : State()
-  data class Operation(val left: Float, val action: Action, val displayNumber: DisplayNumber) : State()
+  data class Operation(val left: Double, val action: Action, val displayNumber: DisplayNumber) : State()
   data class Error(val msg: String) : State()
 
   companion object {
